+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Create type renderer: The item type is not present in the *items.xml files [#1531](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1531)
 - Create type renderer: The class for the item type was not found [#1532](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1532)
 - Refresh type renderer: The item type is not present in the *items.xml files [#1533](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1533)
+- Refresh type renderer: The class for the item type was not found [#1534](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1534)
 
 ### `Other` enhancements
 - Migrated to immutable Settings [#1527](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1527)